requirements.txt
streamlit>=1.28.0
crewai>=0.28.8
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
typing-extensions>=4.0.0